refactor auth and tab redux
