+++ src/redux/actions/tabActions.js
@@ -1,0 +1,3 @@
+export const setTab = (tab) => (dispatch) => {
+  dispatch({ type: tab });
+};
