+++ src/redux/reducers/authReducer.js
@@ -1,9 +1,9 @@
 import AUTH_TYPE from '../types/authTypes';
 
 const initialState = {
-  user: {},
-  isAuthenticated: false,
   isLoading: false,
+  loggedIn: false,
+  user: {},
 };
 // const savedUser = JSON.parse(localStorage.getItem('user'));
 // const initialState = savedUser ? {
@@ -14,7 +14,7 @@ const initialState = {
 //   isLoading: false,
 // } : notLoggedInState;
 
-export default (state = initialState, action) => {
+const authReducer = (state = initialState, action) => {
   switch (action.type) {
     case AUTH_TYPE.USER_LOADING:
       return {
@@ -23,23 +23,25 @@ export default (state = initialState, action) => {
       };
     case AUTH_TYPE.USER_LOADED:
       return {
-        user: { ...action.payload },
-        isAuthenticated: true,
         isLoading: false,
+        loggedIn: true,
+        user: { ...action.payload },
       };
     case AUTH_TYPE.USER_LOGOUT:
       return {
-        user: {},
-        isAuthenticated: false,
         isLoading: false,
+        loggedIn: false,
+        user: {},
       };
     case AUTH_TYPE.AUTH_ERROR:
       return {
-        user: {},
-        isAuthenticated: false,
         isLoading: false,
+        loggedIn: false,
+        user: {},
       };
     default:
       return state;
   }
 };
+
+export default authReducer;
