+++ src/router/PrivateRoute.js
@@ -3,6 +3,7 @@ import React from 'react';
 import { Container } from '@material-ui/core';
 import { Route, Redirect } from 'react-router-dom';
 import PropTypes from 'prop-types';
+import { useDispatch } from 'react-redux';
 
 // Import Routes List
 import Routes from './RouteList';
@@ -13,15 +14,22 @@ import Layout from '../Layout';
 // Import Utils
 import { hasAccess, DEFAULT_REDIRECT } from 'utils/roles';
 
+// Import Actions
+import { setUser } from 'redux/actions/authActions';
+
 const PrivateRoute = ({ component: Component, pageAccess, ...rest }) => {
-  const user = localStorage.getItem('user') ? JSON.parse(localStorage.getItem('user')) : null;
-  const authorized = hasAccess(parseInt(user.role), pageAccess);
+  const dispatch = useDispatch();
+  const auth = JSON.parse(localStorage.getItem('user')) || {};
+  const isAuthorized = auth ? hasAccess(parseInt(auth.role), pageAccess) : false;
+
+  if (auth) dispatch(setUser(auth));
+
   return (
     <div>
       <Route
         {...rest}
         render={(props) => {
-          if (user && authorized) {
+          if (auth && isAuthorized) {
             return (
               <Layout>
                 <Container className={rest.path === '/manage' ? 'custom-container' : ''} maxWidth="md" fixed>
@@ -30,7 +38,7 @@ const PrivateRoute = ({ component: Component, pageAccess, ...rest }) => {
               </Layout>
             );
           }
-          if (user) return <Redirect to={DEFAULT_REDIRECT[user.role]} />;
+          if (isAuthenticated) return <Redirect to={DEFAULT_REDIRECT[user.role]} />;
           else return <Redirect to={Routes.login.root} />;
         }}
       />
